Add type cast

+++ procedures/dither_area_to_not_bigger_than_100pc_of_hex_area.sql
@@ -7,7 +7,7 @@ $$
         carry     jsonb;
         carry_out jsonb;
     begin
-        res = :start_resolution;
+        res = :start_resolution::integer;
         while res >= 0
             loop
                 select jsonb_object_agg(column_name, 0) from unnest('columns_list'::text[]) "column_name" into carry;
